requirements.txt
Flask==2.3.3  # Ajusta la versión según sea necesario
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7 # Por seguridad y compatibilidad con Flask
bcrypt==4.0.1
python-dotenv==1.0.0
python-docx==0.8.11 # Para exportar a Word
Pillow==10.0.0 # Para manejo de imágenes
Flask-Migrate==4.0.4 # Opcional, para migraciones de base de datos con Alembic
email_validator==2.0.0 # Para validación de email en WTForms
# Para Cloudflare Workers (esto podría necesitar ajustes específicos)
# No hay una dependencia directa de Python para Cloudflare Workers en requirements.txt
# La integración se hace a través de Wrangler y adaptando el código Flask.
# Considerar `gunicorn` si se despliega en un entorno WSGI tradicional antes de Workers.
# gunicorn==21.2.0
